Refactored, still have teleporting bug
